+++ sdk/datafactory/azure-resourcemanager-datafactory/src/main/java/com/azure/resourcemanager/datafactory/models/DynamicsLinkedService.java
@@ -28,7 +28,7 @@ public class DynamicsLinkedService extends LinkedService {
      * string (or Expression with resultType string).
      */
     @JsonProperty(value = "typeProperties.deploymentType", required = true)
-    private DynamicsDeploymentType deploymentType;
+    private Object deploymentType;
 
     /*
      * The host name of the on-premises Dynamics server. The property is
@@ -70,7 +70,7 @@ public class DynamicsLinkedService extends LinkedService {
      * scenario. Type: string (or Expression with resultType string).
      */
     @JsonProperty(value = "typeProperties.authenticationType", required = true)
-    private DynamicsAuthenticationType authenticationType;
+    private Object authenticationType;
 
     /*
      * User name to access the Dynamics instance. Type: string (or Expression
@@ -100,7 +100,7 @@ public class DynamicsLinkedService extends LinkedService {
      * resultType string).
      */
     @JsonProperty(value = "typeProperties.servicePrincipalCredentialType")
-    private DynamicsServicePrincipalCredentialType servicePrincipalCredentialType;
+    private Object servicePrincipalCredentialType;
 
     /*
      * The credential of the service principal object in Azure Active
@@ -127,7 +127,7 @@ public class DynamicsLinkedService extends LinkedService {
      *
      * @return the deploymentType value.
      */
-    public DynamicsDeploymentType deploymentType() {
+    public Object deploymentType() {
         return this.deploymentType;
     }
 
@@ -138,7 +138,7 @@ public class DynamicsLinkedService extends LinkedService {
      * @param deploymentType the deploymentType value to set.
      * @return the DynamicsLinkedService object itself.
      */
-    public DynamicsLinkedService withDeploymentType(DynamicsDeploymentType deploymentType) {
+    public DynamicsLinkedService withDeploymentType(Object deploymentType) {
         this.deploymentType = deploymentType;
         return this;
     }
@@ -240,7 +240,7 @@ public class DynamicsLinkedService extends LinkedService {
      *
      * @return the authenticationType value.
      */
-    public DynamicsAuthenticationType authenticationType() {
+    public Object authenticationType() {
         return this.authenticationType;
     }
 
@@ -252,7 +252,7 @@ public class DynamicsLinkedService extends LinkedService {
      * @param authenticationType the authenticationType value to set.
      * @return the DynamicsLinkedService object itself.
      */
-    public DynamicsLinkedService withAuthenticationType(DynamicsAuthenticationType authenticationType) {
+    public DynamicsLinkedService withAuthenticationType(Object authenticationType) {
         this.authenticationType = authenticationType;
         return this;
     }
@@ -328,7 +328,7 @@ public class DynamicsLinkedService extends LinkedService {
      *
      * @return the servicePrincipalCredentialType value.
      */
-    public DynamicsServicePrincipalCredentialType servicePrincipalCredentialType() {
+    public Object servicePrincipalCredentialType() {
         return this.servicePrincipalCredentialType;
     }
 
@@ -340,8 +340,7 @@ public class DynamicsLinkedService extends LinkedService {
      * @param servicePrincipalCredentialType the servicePrincipalCredentialType value to set.
      * @return the DynamicsLinkedService object itself.
      */
-    public DynamicsLinkedService withServicePrincipalCredentialType(
-        DynamicsServicePrincipalCredentialType servicePrincipalCredentialType) {
+    public DynamicsLinkedService withServicePrincipalCredentialType(Object servicePrincipalCredentialType) {
         this.servicePrincipalCredentialType = servicePrincipalCredentialType;
         return this;
     }