+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/main/java/com/azure/ai/metricsadvisor/models/MetricsAdvisorKeyCredential.java
@@ -3,12 +3,15 @@
 
 package com.azure.ai.metricsadvisor.models;
 
+import com.azure.core.annotation.Fluent;
+
 /**
  * The MetricsAdvisorKeyCredential class.
  */
+@Fluent
 public final class MetricsAdvisorKeyCredential {
-    private volatile String subscriptionKey;
-    private volatile String apiKey;
+    private volatile MetricsAdvisorKeys keys;
+    private final Object updateLock = new Object();
 
     /**
      * Creates a MetricsAdvisorKeyCredential credential that authorizes request with the given keys.
@@ -17,54 +20,33 @@ public final class MetricsAdvisorKeyCredential {
      * @param apiKey the api key used to authorize requests
      */
     public MetricsAdvisorKeyCredential(String subscriptionKey, String apiKey) {
-        this.subscriptionKey = subscriptionKey;
-        this.apiKey = apiKey;
+        this.keys = new MetricsAdvisorKeys(subscriptionKey, apiKey);
     }
 
     /**
-     * Retrieves the subscription key associated to this credential.
+     * Retrieves the {@link MetricsAdvisorKeys} containing the subscription key and api key
+     * associated with this credential.
      *
-     * @return The subscription key being used to authorize requests.
+     * @return The {@link MetricsAdvisorKeys} containing the subscription key and api key.
      */
-    public String getSubscriptionKey() {
-        return this.subscriptionKey;
+    public MetricsAdvisorKeys getKeys() {
+        return this.keys;
     }
 
     /**
-     * Retrieves the api key associated to this credential.
-     *
-     * @return The api key being used to authorize requests.
-     */
-    public String getApiKey() {
-        return this.apiKey;
-    }
-
-    /**
-     * Rotates the subscription key associated to this credential.
+     * Update the subscription and api key associated to this credential.
      * <p>
-     * This is intended to be used when you've regenerated your subscription key and want to
-     * update long lived clients.
+     * This is intended to be used when you've regenerated your subscription key and
+     * api key want to update long lived clients.
      * </p>
      * @param subscriptionKey The new subscription key to associated with this credential.
-     * @return The updated {@code MetricsAdvisorKeyCredential} object.
-     */
-    public MetricsAdvisorKeyCredential updateSubscriptionKey(String subscriptionKey) {
-        this.subscriptionKey = subscriptionKey;
-        return this;
-    }
-
-    /**
-     * Rotates the api key associated to this credential.
-     * <p>
-     * This is intended to be used when you've regenerated your api key and want to
-     * update long lived clients.
-     * </p>
-     *
      * @param apiKey The new api key to associated with this credential.
      * @return The updated {@code MetricsAdvisorKeyCredential} object.
      */
-    public MetricsAdvisorKeyCredential updateApiKey(String apiKey) {
-        this.apiKey = apiKey;
+    public MetricsAdvisorKeyCredential updateKey(String subscriptionKey, String apiKey) {
+        synchronized (this.updateLock) {
+            this.keys = new MetricsAdvisorKeys(subscriptionKey, apiKey);
+        }
         return this;
     }
 }