+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/test/java/com/azure/ai/metricsadvisor/DetectionConfigurationTest.java
@@ -4,11 +4,9 @@
 package com.azure.ai.metricsadvisor;
 
 import com.azure.ai.metricsadvisor.administration.MetricsAdvisorAdministrationClient;
-import com.azure.ai.metricsadvisor.models.AnomalyDetectionConfiguration;
-import com.azure.ai.metricsadvisor.models.DataFeed;
-import com.azure.ai.metricsadvisor.models.DataFeedMetric;
-import com.azure.ai.metricsadvisor.models.ListMetricAnomalyDetectionConfigsOptions;
-import com.azure.ai.metricsadvisor.models.MetricsAdvisorServiceVersion;
+import com.azure.ai.metricsadvisor.administration.models.AnomalyDetectionConfiguration;
+import com.azure.ai.metricsadvisor.administration.models.DataFeed;
+import com.azure.ai.metricsadvisor.administration.models.DataFeedMetric;
 import com.azure.core.http.HttpClient;
 import com.azure.core.test.TestBase;
 import com.azure.core.util.CoreUtils;
@@ -143,7 +141,7 @@ public class DetectionConfigurationTest extends DetectionConfigurationTestBase {
             assertNotNull(configuration);
             id.set(configuration.getId());
 
-            client.listMetricAnomalyDetectionConfigs(costMetricId, new ListMetricAnomalyDetectionConfigsOptions())
+            client.listMetricAnomalyDetectionConfigs(costMetricId)
                 .forEach(config -> Assertions.assertNotNull(config));
         } finally {
             if (!CoreUtils.isNullOrEmpty(id.get())) {