+++ sdk/communication/azure-communication-identity/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.2.0-beta.1 (Unreleased)
 
+
+## 1.1.1 (2021-06-09)
+Updated `azure-communication-identity` version
+
 ## 1.1.0 (2021-05-27)
 ### Added
 - New exception `IdentityErrorResponseException` was added for more descriptive server error messages.