+++ sdk/videoanalyzer/azure-media-videoanalyzer-edge/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 1.0.0-beta.4 (2021-06-10)
+
+- Update the description of the package, to add more info on the maven page.
+- Update dependency version, `azure-core` to 1.17.0
+
 ## 1.0.0-beta.3 (2021-05-24)
 
 - Updated the title and description of the package and updates to readme.